Rename AlternativeID page to AlternativeApplicantID
